+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "altaqwaa",
-  "version": "2.1.0",
+  "version": "3.0.0",
   "homepage": "https://www.altaqwaa.org",
   "description": "Altaqwaa - Islamic Desktop Application",
   "main": "./src/index.js",
