+++ src/pages/settings.html
@@ -140,6 +140,33 @@
                     </td>
                 </tr>
                 <tr>
+                    <td class="name">بدأ التطبيق مع النظام (لويندوز فقط)</td>
+                    <td>
+                        <label class="switch" id="label_adhkar">
+                            <input type="checkbox" id="autostart">
+                            <span class="slider round"></span>
+                        </label>
+                    </td>
+                </tr>
+                <tr>
+                    <td class="name">بدأ التطبيق بالخلفية (بدون واجهة)</td>
+                    <td>
+                        <label class="switch" id="label_adhkar">
+                            <input type="checkbox" id="startHidden">
+                            <span class="slider round"></span>
+                        </label>
+                    </td>
+                </tr>
+                <tr>
+                    <td class="name">تصغير التطبيق لشريط المهام (تلقائيا)</td>
+                    <td>
+                        <label class="switch" id="label_adhkar">
+                            <input type="checkbox" id="minimizeToPanel">
+                            <span class="slider round"></span>
+                        </label>
+                    </td>
+                </tr>
+                <tr>
                     <td class="name">مستوى الصوت (للإستماع للقرآن)</td>
                     <td>
                         <input type="range" value="100" min="1" max="100" id="volume">
