+++ src/window/audio_window.js
@@ -39,7 +39,9 @@ module.exports = function audio_window(path, App_Path, BrowserWindow, ipcMain, a
     win?.loadFile(path.join(__dirname, '../pages/audio_window.html'));
 
     win?.once('ready-to-show', () => {
+        console.log("[Altaqwaa-CLI] Audio Window Ready.")
         win?.hide();
+        console.log("[Altaqwaa-CLI] Audio Window Hidden.")
     });
 
     win?.on('closed', (event) => {
